Move `Cas1ApplicationUpdateFields`
This was missed in a previous commit and should have been moved to `Cas1ApplicationCreationService`
